pyproject.toml
[tool.poetry]
name = "updater"
version = "0.1.0"
description = ""
authors = ["Vladimir Pouzanov <farcaller@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.7.0"
python-dxf = "^10.0.0"
requests = "^2.28.2"
chevron = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
updater = "updater:app"
